refactor(controller): using messaging controller
